refactor: スライダー値型を usize から i32 へ変更
PanelEvent::SetValue / DragValue および PanelNode::Slider の
min / max / value / display_value を i32 に統一。
負のスライダー値（例: 回転角度）を将来的に扱えるようにする。

+++ apps/desktop/src/app/panel_dispatch.rs
@@ -10,7 +10,7 @@ pub(crate) enum PanelDragState {
     Control {
         panel_id: String,
         node_id: String,
-        source_value: usize,
+        source_value: i32,
     },
     Move {
         panel_id: String,
